+++ api/v1alpha1/nifiparametercontext_types.go
@@ -17,6 +17,8 @@ type NifiParameterContextSpec struct {
 	ClusterRef ClusterReference `json:"clusterRef,omitempty"`
 	// a list of secret containing sensitive parameters (the key will name of the parameter).
 	SecretRefs []SecretReference `json:"secretRefs,omitempty"`
+	// whether or not the operator should take over an existing parameter context if its name is the same.
+	DisableTakeOver *bool `json:"disableTakeOver,omitempty"`
 }
 
 type Parameter struct {
@@ -82,4 +84,11 @@ type NifiParameterContextList struct {
 
 func init() {
 	SchemeBuilder.Register(&NifiParameterContext{}, &NifiParameterContextList{})
+}
+
+func (d *NifiParameterContextSpec) IsTakeOverEnabled() bool {
+	if d.DisableTakeOver == nil {
+		return true
+	}
+	return !*d.DisableTakeOver
 }